+++ native/cocos/renderer/core/gfx/GFXQueue.h
@@ -17,7 +17,7 @@ public:
     virtual void submit(const CommandBuffer *const *cmdBuffs, uint count, Fence *fence) = 0;
 
     CC_INLINE void submit(const vector<CommandBuffer *> &cmdBuffs) { submit(cmdBuffs, nullptr); }
-    CC_INLINE void submit(const vector<CommandBuffer *> &cmdBuffs, Fence *fence) { submit(cmdBuffs.data(), cmdBuffs.size(), fence); }
+    CC_INLINE void submit(const vector<CommandBuffer *> &cmdBuffs, Fence *fence) { submit(cmdBuffs.data(), static_cast<uint>(cmdBuffs.size()), fence); }
     CC_INLINE Device *getDevice() const { return _device; }
     CC_INLINE QueueType getType() const { return _type; }
     CC_INLINE bool isAsync() const { return _isAsync; }
